# zwf-app/src/_assets/api.yml
swagger: "2.0"
info:
  description: ZWF backend API
  version: "1.0.0"
  title: "ZWF API"
basePath: "/api/v1"
schemes:
  - http
  - https
...
paths:
...
  /user/:id:
    delete:
      operationId: deleteUser
      # security:
      #   - authAdminOrAgent: []
      parameters:
        - name: id
          in: path
          type: string
          format: uuid
          required: true
      responses:
        200:
          description: OK
...
  /user/:id/set_password:
    post:
      operationId: setUserPassword
      # security:
      #   - authAdminOrAgent: []
      parameters:
        - name: id
          in: path
          type: string
          format: uuid
          required: true
      responses:
        200:
          description: OK
...
  /user/:id/tags:
    post:
      operationId: setUserTags
      parameters:
        - name: id
          in: path
          type: string
          format: uuid
          required: true
      responses:
        200:
          description: OK
  /user/:id/role:
    post:
      operationId: setUserRole
      parameters:
        - name: id
          in: path
          type: string
          format: uuid
          required: true
      responses:
        200:
          description: OK
...
  /task_template/{id}:
    get:
      operationId: getTaskTemplate
      # security:
      #   - authLoggedInUser: []
      parameters:
        - name: id
          in: path
          type: string
          format: uuid
          required: true
      responses:
        200:
          description: OK
    delete:
      operationId: deleteTaskTemplate
      # security:
      #   - authAdmin: []
      parameters:
        - name: id
          in: path
          type: string
          format: uuid
          required: true
      responses:
        200:
          description: OK
  /task_template/{id}/clone:
    put:
      operationId: cloneTaskTemplate
      # security:
      #   - authLoggedInUser: []
      parameters:
        - name: id
          in: path
          type: string
          format: uuid
          required: true
      responses:
        200:
          description: OK
  /task_template/{id}/rename:
    post:
      operationId: renameTaskTemplate
      # security:
      #   - authClient: []
      parameters:
        - name: id
          in: path
          type: string
          format: uuid
          required: true
      responses:
        200:
          description: OK
...
  /task/{id}/tracking:
    get:
      operationId: listTaskTrackings
      parameters:
        - name: id
          in: path
          type: string
          format: uuid
          required: true
      responses:
        200:
          description: OK
    post:
      operationId: createNewTaskTracking
      parameters:
        - name: id
          in: path
          type: string
          format: uuid
          required: true
      responses:
        200:
          description: OK
...
  /task/{id}/tracking/sse:
    get:
      operationId: subscribeTaskTracking
      parameters:
        - name: id
          in: path
          type: string
          format: uuid
          required: true
      responses:
        200:
          description: OK
  /task/{id}/tracking/nudge:
    get:
      operationId: nudgeTrackingCursor
      parameters:
        - name: taskId
          in: path
          type: string
          format: uuid
          required: true
      responses:
        200:
          description: OK
...